# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jewelry API",
        "description": "API for Jewelry Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/financial-management/fees": {
            "get": {
                "tags": [
                    "Fees"
                ],
                "summary": "Get all fees",
                "description": "Retrieve a paginated list of all fees",
                "operationId": "2de7020cadb8857fb29ca4b27336496c",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fees retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 100
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Fees"
                ],
                "summary": "Create a new fee",
                "description": "Create a new fee configuration",
                "operationId": "0fd953ad2e3f3ec1ae5c0af3a9cd2718",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "amount",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Processing Fee"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "fixed",
                                            "percentage"
                                        ],
                                        "example": "fixed"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Processing fee for all transactions"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Fee created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fee": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/financial-management/fees/{id}": {
            "get": {
                "tags": [
                    "Fees"
                ],
                "summary": "Get a specific fee",
                "description": "Retrieve detailed information about a specific fee",
                "operationId": "ec62de9e4d06fdbcd6d9e756b3d26ccf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Fee ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fee retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fee": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Fee not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Fees"
                ],
                "summary": "Update a fee",
                "description": "Update an existing fee configuration",
                "operationId": "feaba0857b015878834eeac3168673a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Fee ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "amount",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Processing Fee"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 30
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "fixed",
                                            "percentage"
                                        ],
                                        "example": "fixed"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated processing fee"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fee updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fee": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Fee not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Fees"
                ],
                "summary": "Delete a fee",
                "description": "Move a fee to trash (soft delete)",
                "operationId": "463d4b93517d32048e20105ad60adad1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Fee ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fee moved to trash successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource has been moved to trash."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Fee not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/options/fees": {
            "get": {
                "tags": [
                    "Fees"
                ],
                "summary": "Get active fees for dropdown",
                "description": "Retrieve a list of active fees for dropdown/select options",
                "operationId": "036b0b6f11bdb5ab5decedcc434be968",
                "responses": {
                    "200": {
                        "description": "Active fees retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fees": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Processing Fee"
                                                    },
                                                    "amount": {
                                                        "type": "number",
                                                        "example": 25
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "fixed"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}